[Chore] estatico-webpack: Added default config files for webpack, babel and browserslist
