+++ packages/estatico-boilerplate/.babelrc.js
@@ -1,0 +1,3 @@
+const defaults = require('@unic/estatico-webpack/.babelrc.js');
+
+module.exports = defaults;
