+++ packages/estatico-boilerplate/webpack.config.js
@@ -1,0 +1,70 @@
+const defaults = require('@unic/estatico-webpack/webpack.config.js');
+const env = require('minimist')(process.argv.slice(2));
+const merge = require('lodash.merge');
+const glob = require('glob');
+const path = require('path');
+
+module.exports = [
+  merge({}, defaults, {
+    entry: Object.assign({
+      head: './src/assets/js/head.js',
+      main: './src/assets/js/main.js',
+    }, (env.dev || env.ci) ? {
+      dev: './src/assets/js/dev.js',
+    } : {}),
+    output: {
+      path: path.resolve('./dist/assets/js'),
+      filename: `[name]${env.dev ? '' : '.min'}.js`,
+      chunkFilename: `async/[name]${env.dev ? '' : '.min'}.js`,
+      publicPath: '/assets/js/',
+    },
+    mode: env.dev ? 'development' : 'production',
+  }),
+  {
+    entry: {
+      test: './src/preview/assets/js/test.js',
+    },
+    module: {
+      rules: defaults.module.rules.concat([
+        {
+          test: /qunit\.js$/,
+          loader: 'expose-loader?QUnit',
+        },
+        {
+          test: /\.css$/,
+          loader: 'style-loader!css-loader',
+        },
+      ]),
+    },
+    externals: {
+      jquery: 'jQuery',
+    },
+    output: {
+      path: path.resolve('./dist/preview/assets/js'),
+      filename: `[name]${env.dev ? '' : '.min'}.js`,
+      chunkFilename: `async/[name]${env.dev ? '' : '.min'}.js`,
+    },
+    mode: 'development',
+  },
+  {
+    // Create object of fileName:filePath pairs
+    entry: glob.sync('./src/**/*.test.js').reduce((obj, item) => {
+      const key = path.basename(item, path.extname(item));
+
+      obj[key] = item; // eslint-disable-line no-param-reassign
+
+      return obj;
+    }, {}),
+    module: defaults.module,
+    externals: {
+      jquery: 'jQuery',
+      qunit: 'QUnit',
+    },
+    output: {
+      path: path.resolve('./dist/preview/assets/js/test'),
+      filename: `[name]${env.dev ? '' : '.min'}.js`,
+      chunkFilename: `async/[name]${env.dev ? '' : '.min'}.js`,
+    },
+    mode: 'development',
+  },
+];
